Add address prefixes and addressing methods
Add address prefixes as consts and addressing functions to the protocol
module. These will be used to convert a Sabre payload builder into a
Transact transaction builder. Additionally, these are made publicly
available so consumers of the API can properly calculate addresses.

The `transact` crate is pulled in as a dependency only when the target
is not `wasm32`; this is because Transact's cbor dependency is not
compatible with WASM.

Signed-off-by: Logan Seeley <seeley@bitwise.io>